feat: JSDoc parity, TS erased syntax, conformance V4 TS oracle
Implement the full JSDoc surface (TDD-00125): @param/@returns/@typedef/
@callback/@template, the Closure type-expression grammar (Array.<T>,
function(...), ?T/!T, unions, object shapes, import("...") types), and
the class/visibility tags. Add integer-width @type/@param overrides
(TDD-00123) and the TS erased-syntax pass (TDD-00124). Extend the
conformance runner with the TypeScript accept/reject oracle, and refresh
the docs website (install guide, examples, coverage).

+++ codegen/llvm/emit_call_encoding.go
@@ -79,7 +79,7 @@ func (e *Emitter) emitCryptoGetRandomValues(args []ast.Expression, pos ast.Pos) 
 	if !ok {
 		return Value{}, fmt.Errorf("%d:%d: undefined variable '%s'", pos.Line, pos.Col, id.Name)
 	}
-	if !sym.Ty.IsArray || sym.Ty.ElemType == nil || sym.Ty.ElemType.IR != "i64" {
+	if !sym.Ty.IsArray || sym.Ty.ElemType == nil || sym.Ty.ElemType.IR != "double" {
 		return Value{}, fmt.Errorf("%d:%d: crypto.getRandomValues requires a TypedArray, an ArrayBuffer, or a number[] array", pos.Line, pos.Col)
 	}
